requirements.txt
# Django framework
Django==5.2.5

# Django REST Framework
djangorestframework==3.16.1

# JWT authentication
djangorestframework-simplejwt==5.2.1

# PostgreSQL driver
psycopg2-binary==2.9.7

# Required by Django
asgiref==3.9.1
pytz==2025.2
sqlparse==0.5.3

# Optional: For handling images if you use ImageField
Pillow==10.1.0

# CORS headers
django-cors-headers==4.4.0

# Environment variables
python-decouple==3.8
